List Github projects for better usability
