+++ cmd/droned/assets/css/drone.less
@@ -184,7 +184,7 @@ body {
   border:0px;
 }
 
-// 
+//
 // nav-repos
 // --------------------------------------------
 
@@ -364,7 +364,7 @@ body {
 	margin-bottom: 10px;
 }
 
-// 
+//
 // build list
 // --------------------------------------------
 
@@ -493,7 +493,7 @@ body {
 .btn.btn-Success {
 	background:rgba(81, 163, 81, 0.75);
 }
-.btn.btn-failure, 
+.btn.btn-failure,
 .btn.btn-Failure,
 .btn.btn-Error {
 	background:rgba(189, 54, 47, 0.8);
@@ -911,7 +911,7 @@ textarea {
 	top: 6px;
 	transition: all .15s ease;
 	width: 35px;
-	z-index: 3; 
+	z-index: 3;
 	border-radius:5px;
 }
 
@@ -1284,4 +1284,21 @@ pre {
   cursor: pointer;
 }
 
-
+.github-projects {
+	.spinner {
+		padding-top: 14px;
+		padding-bottom: 10px;
+	}
+	.projects-list {
+		.item {
+			display: inline-block;
+			padding: 2px 3px;
+			margin: 2px 3px;
+			background-color: #F4F4F4;
+			a {
+				font-size: 12px;
+				color: #777;
+			}
+		}
+	}
+}
