+++ cmd/droned/assets/js/github_repos.js
@@ -1,0 +1,8 @@
+GithubRepos = {
+  url: "/new/github.com/available_repos",
+  get: function(success) {
+    $.getJSON(this.url, function(response) {
+      success(response)
+    })
+  }
+}
