+++ cmd/droned/assets/js/spin.js
@@ -1,253 +1,3 @@
-;// Live commit updates
-
-if(typeof(Drone) === 'undefined') { Drone = {}; }
-
-(function () {
-	Drone.CommitUpdates = function(socket) {
-		if(typeof(socket) === "string") {
-			var url = [(window.location.protocol == 'https:' ? 'wss' : 'ws'),
-								 '://',
-								 window.location.host,
-								 socket].join('')
-			this.socket = new WebSocket(url);
-		} else {
-			this.socket = socket;
-		}
-
-		this.lineFormatter = new Drone.LineFormatter();
-		this.attach();
-	}
-
-	Drone.CommitUpdates.prototype = {
-		lineBuffer: "",
-		autoFollow: false,
-
-		startOutput: function(el) {
-			if(typeof(el) === 'string') {
-				this.el = document.getElementById(el);
-			} else {
-				this.el = el;
-			}
-
-			if(!this.reqId) {
-				this.updateScreen();
-			}
-		},
-
-		stopOutput: function() {
-			this.stoppingRefresh = true;
-		},
-
-		attach: function() {
-			this.socket.onopen    = this.onOpen;
-			this.socket.onerror   = this.onError;
-			this.socket.onmessage = this.onMessage.bind(this);
-			this.socket.onclose   = this.onClose;
-		},
-
-		updateScreen: function() {
-			if(this.lineBuffer.length > 0) {
-				this.el.innerHTML += this.lineBuffer;
-				this.lineBuffer = '';
-
-				if (this.autoFollow) {
-					window.scrollTo(0, document.body.scrollHeight);
-				}
-			}
-
-			if(this.stoppingRefresh) {
-				this.stoppingRefresh = false;
-			} else {
-				window.requestAnimationFrame(this.updateScreen.bind(this));
-			}
-		},
-
-		onOpen: function() {
-			console.log('output websocket open');
-		},
-
-		onError: function(e) {
-			console.log('websocket error: ' + e);
-		},
-
-		onMessage: function(e) {
-			this.lineBuffer += this.lineFormatter.format(e.data);
-		},
-
-		onClose: function(e) {
-			console.log('output websocket closed: ' + JSON.stringify(e));
-			window.location.reload();
-		}
-	};
-
-	// Polyfill rAF for older browsers
-	window.requestAnimationFrame = window.requestAnimationFrame ||
-		window.webkitRequestAnimationFrame ||
-		function(callback, element) {
-			return window.setTimeout(function() {
-				callback(+new Date());
-			}, 1000 / 60);
-		};
-
-	window.cancelRequestAnimationFrame = window.cancelRequestAnimationFrame ||
-		window.cancelWebkitRequestAnimationFrame ||
-		function(fn) {
-			window.clearTimeout(fn);
-		};
-
-})();
-$(function() {
-  var projectForm = $(".form-repo")
-  var repoOwnerField = projectForm.find("input[name=owner]")
-  var repoNameField = projectForm.find("input[name=name]")
-
-  var projectsBlock = $(".github-projects")
-  var projectsList = projectsBlock.find(".projects-list")
-  var spinnerBlock = projectsBlock.find(".spinner")
-
-  spinner = new Spinner({
-    lines: 12,
-    speed: 0.5,
-    width: 4,
-    length: 10
-  })
-  spinner.spin(spinnerBlock[0])
-
-  GithubRepos.get(function(response) {
-    $.each(response, function(i, repo) {
-      var title = repo.owner + "/" + repo.name
-
-      item = $("<div></div>").addClass("item")
-      link = $("<a></a>").text(title).attr("href", "#")
-      if(repo.private) {
-        icon = $("<span></span>").addClass("glyphicon").addClass("glyphicon-lock")
-        item.append("&nbsp;")
-        item.append(icon)
-      }
-
-      item.append(link)
-
-      item.data('owner', repo.owner)
-      item.data('name', repo.name)
-
-      projectsList.append(item)
-    })
-
-    spinner.stop()
-    spinnerBlock.remove()
-  })
-
-  projectsList.on('click', 'a', function(event) {
-    var badge = $(event.target).parent()
-
-    repoOwnerField.val(badge.data('owner'))
-    repoNameField.val(badge.data('name'))
-    repoNameField.focus()
-
-    $(document).scrollTop(projectForm.offset().top)
-
-    return false
-  })
-
-  projectForm.on('submit', function() {
-    $("#successAlert").hide();
-    $("#failureAlert").hide();
-    $('#submitButton').button('loading');
-
-    $.ajax({
-      type: "POST",
-      url: projectForm.attr("target"),
-      data: projectForm.serialize(),
-      success: function(response, status) {
-        var name = $("input[name=name]").val()
-        var owner = $("input[name=owner]").val()
-        var domain = $("input[name=domain]").val()
-        window.location.pathname = "/" + domain + "/"+owner+"/"+name
-      },
-      error: function() {
-        $("#failureAlert").text("Unable to setup the Repository");
-        $("#failureAlert").show().removeClass("hide");
-        $('#submitButton').button('reset');
-      }
-    });
-
-    return false;
-  })
-})
-GithubRepos = {
-  url: "/new/github.com/available_repos",
-  get: function(success) {
-    $.getJSON(this.url, function(response) {
-      success(response)
-    })
-  }
-}
-;// Format ANSI to HTML
-
-if(typeof(Drone) === 'undefined') { Drone = {}; }
-
-(function() {
-	Drone.LineFormatter = function() {};
-
-	Drone.LineFormatter.prototype = {
-		regex: /\u001B\[([0-9]+;?)*[Km]/g,
-		styles: [],
-
-		format: function(s) {
-			// Check for newline and early exit?
-			s = s.replace(/</g, "&lt;");
-			s = s.replace(/>/g, "&gt;");
-
-			var output = "";
-			var current = 0;
-			while (m = this.regex.exec(s)) {
-				var part = s.substring(current, m.index);
-				current = this.regex.lastIndex;
-
-				var token = s.substr(m.index, this.regex.lastIndex - m.index);
-				var code = token.substr(2, token.length-2);
-
-				var pre = "";
-				var post = "";
-
-				switch (code) {
-					case 'm':
-					case '0m':
-						var len = this.styles.length;
-						for (var i=0; i < len; i++) {
-							this.styles.pop();
-							post += "</span>"
-						}
-						break;
-					case '30;42m': pre = '<span style="color:black;background:lime">'; break;
-					case '36m':
-					case '36;1m': pre = '<span style="color:cyan;">'; break;
-					case '31m':
-					case '31;31m': pre = '<span style="color:red;">'; break;
-					case '33m':
-					case '33;33m': pre = '<span style="color:yellow;">'; break;
-					case '32m':
-					case '0;32m': pre = '<span style="color:lime;">'; break;
-					case '90m': pre = '<span style="color:gray;">'; break;
-					case 'K':
-					case '0K':
-					case '1K':
-					case '2K': break;
-				}
-
-				if (pre !== "") {
-					this.styles.push(pre);
-				}
-
-				output += part + pre + post;
-			}
-
-			var part = s.substring(current, s.length);
-			output += part;
-			return output;
-		}
-	};
-})();
 /**
  * Copyright (c) 2011-2013 Felix Gnass
  * Licensed under the MIT license
