+++ cmd/droned/drone.go
@@ -156,6 +156,7 @@ func setupHandlers() {
 	// handlers for setting up your GitHub repository
 	m.Post("/new/github.com", handler.UserHandler(handler.RepoCreateGithub))
 	m.Get("/new/github.com", handler.UserHandler(handler.RepoAdd))
+	m.Get("/new/github.com/available_repos", handler.UserHandler(handler.AvailableGithubRepos))
 
 	// handlers for linking your GitHub account
 	m.Get("/auth/login/github", handler.UserHandler(handler.LinkGithub))
