+++ pkg/template/pages/base.html
@@ -19,8 +19,6 @@
 		<link href="//fonts.googleapis.com/css?family=Open&#43;Sans" rel="stylesheet" />
 		<link href="//fonts.googleapis.com/css?family=Droid&#43;Sans&#43;Mono" rel="stylesheet" />
 		<link href="//cdnjs.cloudflare.com/ajax/libs/font-awesome/4.0.3/css/font-awesome.min.css" rel="stylesheet" />
-
-		<script src="/js/main.js" type="text/javascript"></script>
 	</head>
 	<body>
 
@@ -58,6 +56,7 @@
 		{{ template "content" . }}
 		<script src="//cdnjs.cloudflare.com/ajax/libs/jquery/2.0.3/jquery.min.js"></script>
 		<script src="//cdnjs.cloudflare.com/ajax/libs/twitter-bootstrap/3.0.3/js/bootstrap.min.js"></script>
+		<script src="/js/main.js" type="text/javascript"></script>
 		{{ template "script" . }}
 	</body>
 </html>
