+++ pkg/template/pages/github_add.html
@@ -43,6 +43,15 @@
 								</div>
 							</div>
 						</div>
+
+						<div class="alert">
+							Select from existing Github repos
+						</div>
+						<div class="github-projects">
+							<div class="spinner"></div>
+							<div class="projects-list"></div>
+						</div>
+
 						<br/>
 						<div class="alert">Select your Drone account</div>
 						<ul>
@@ -72,30 +81,4 @@
 {{ end }}
 
 {{ define "script" }}
-	<script>
-		document.forms[0].onsubmit = function(event) {
-			$("#successAlert").hide();
-			$("#failureAlert").hide();
-			$('#submitButton').button('loading')
-			
-			var form = event.target
-			var formData = new FormData(form);
-			xhr = new XMLHttpRequest();
-			xhr.open('POST', form.action);
-			xhr.onload = function() {
-				if (this.status == 200) {
-					var name = $("input[name=name]").val()
-					var owner = $("input[name=owner]").val()
-					var domain = $("input[name=domain]").val()
-					window.location.pathname = "/" + domain + "/"+owner+"/"+name
-				} else {
-					$("#failureAlert").text("Unable to setup the Repository");
-					$("#failureAlert").show().removeClass("hide");
-					$('#submitButton').button('reset')
-				};
-			};
-			xhr.send(formData);
-			return false;
-		}
-	</script>
 {{ end }}
